Normalize partner age range ordering
Agent-Logs-Url: https://github.com/marriagestationpvtltd-lang/maapk/sessions/b29bd465-15c2-468e-9217-dc6b7c156053

+++ msfinal/lib/service/partner_age_preferences.dart
@@ -35,8 +35,7 @@ PartnerAgePreferenceBounds resolvePartnerAgePreferenceBounds({
       ? (fallbackMaxAge < minAge ? minAge : fallbackMaxAge)
       : ((_calculateAge(parsedBirthDate) +
                   PartnerAgePreferenceBounds.maximumAgeOffsetFromUserAge)
-              .clamp(minAge, PartnerAgePreferenceBounds.absoluteMaximumAge)
-          as int);
+              .clamp(minAge, PartnerAgePreferenceBounds.absoluteMaximumAge));
 
   return PartnerAgePreferenceBounds(
     minAge: minAge,
